added a separate layer for testing media player
